Face_Login_Project-main/face_login_project/project/requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.10.1
Pillow==10.0.1
supabase==2.3.0
python-dotenv==1.0.0
imutils==0.5.4
matplotlib==3.7.2